requirements.txt
argon2-cffi==23.1.0
pynacl==1.5.0
fastapi==0.115.2
uvicorn==0.30.5
requests==2.31.0
pydantic==2.8.2
cryptography==42.0.5
python-multipart==0.0.9
pyotp==2.9.0
qrcode[pil]==7.4.2